# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MyBlog API
  version: 1.0.0
  description: Vue 前端已验证的数据契约。评论接口由 Artalk 独立提供。
servers:
  - url: /api
...
paths:
...
  /admin/posts:
    get:
      tags: [Admin Posts]
      summary: 获取管理文章列表
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PostStatus' }
      responses:
        '200':
          description: 包括草稿和未来文章的分页数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostPageResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Admin Posts]
      summary: 创建文章
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostMutation' }
      responses:
        '201':
          description: 创建成功
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostDetailResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /admin/administrators:
    get:
      tags: [Admin Permissions]
      summary: 列出管理员（仅站点所有者）
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: 管理员列表
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdministratorListResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin Permissions]
      summary: 按 GitHub 数字 ID 添加管理员（仅站点所有者）
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [githubId]
              properties:
                githubId: { type: integer, format: int64, minimum: 1 }
      responses:
        '201':
          description: 授权成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdministratorResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: blog_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
...
  responses:
    BadRequest:
      description: 请求参数或字段错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 未登录或会话失效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 权限或 CSRF 校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Slug 冲突、Slug 锁定、分类仍被使用、媒体被引用或状态冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    PostStatus:
      type: string
      enum: [draft, published, scheduled]
...
    Tag:
      type: object
      required: [id, name, slug, postCount]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 80 }
        slug: { type: string, maxLength: 80 }
        postCount: { type: integer, minimum: 0 }
    Category:
      type: object
      required: [id, name, slug, description, postCount]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string, maxLength: 80 }
        slug: { type: string, maxLength: 80 }
        description: { type: [string, 'null'], maxLength: 300 }
        postCount: { type: integer, minimum: 0 }
    PostLink:
      type: object
      required: [title, slug]
      properties:
        title: { type: string }
        slug: { type: string }
    PostSummary:
      type: object
      required: [id, title, slug, excerpt, coverUrl, status, publishedAt, updatedAt, category, tags, wordCount, readingTimeMinutes]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, maxLength: 200 }
        slug: { type: string, maxLength: 160 }
        excerpt: { type: string, maxLength: 500 }
        coverUrl: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/PostStatus' }
        publishedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
        category:
          oneOf:
            - $ref: '#/components/schemas/Category'
            - type: 'null'
        tags:
          type: array
          items: { $ref: '#/components/schemas/Tag' }
        wordCount: { type: integer, minimum: 0 }
        readingTimeMinutes: { type: integer, minimum: 1 }
    PostDetail:
      allOf:
        - $ref: '#/components/schemas/PostSummary'
        - type: object
          required: [contentMarkdown, previousPost, nextPost]
          properties:
            contentMarkdown: { type: string }
            previousPost:
              oneOf:
                - $ref: '#/components/schemas/PostLink'
                - type: 'null'
            nextPost:
              oneOf:
                - $ref: '#/components/schemas/PostLink'
                - type: 'null'
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 1 }
    PostPage:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/PostSummary' }
        pagination: { $ref: '#/components/schemas/Pagination' }
    PostPageResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/PostPage' }
    PostDetailResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/PostDetail' }
    PostMutation:
      type: object
      additionalProperties: false
      required: [title, slug, excerpt, contentMarkdown, coverUrl, status, publishedAt, categoryId, tagIds]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug:
          type: string
          minLength: 1
          maxLength: 160
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        excerpt: { type: string, maxLength: 500 }
        contentMarkdown: { type: string, minLength: 1, maxLength: 2097152 }
        coverUrl: { type: [string, 'null'], maxLength: 2048 }
        status: { $ref: '#/components/schemas/PostStatus' }
        publishedAt: { type: [string, 'null'], format: date-time }
        categoryId: { type: [integer, 'null'], format: int64 }
        tagIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items: { type: integer, format: int64, minimum: 1 }
...
    Administrator:
      type: object
      required: [githubId, isOwner, grantedAt]
      properties:
        githubId: { type: integer, format: int64, minimum: 1 }
        isOwner: { type: boolean }
        grantedAt: { type: [string, 'null'], format: date-time }
    AdministratorResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Administrator' }
    AdministratorListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Administrator' }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        requestId: { type: string }